move upload / create lib to utils; add to train_km
