+++ projects/modular_llm/train_experts.py
@@ -7,7 +7,6 @@ from pytorch_lightning import Trainer, seed_everything
 from mttl.arguments import Args, ExpertConfig
 from mttl.datamodule.base import get_datamodule
 from mttl.logging import logger, setup_logging
-from mttl.models.library.expert_library import ExpertLibrary
 from mttl.models.lightning.callbacks import (
     DownstreamEvalCallback,
     LiveCheckpointCallback,
@@ -17,7 +16,12 @@ from mttl.models.lightning.callbacks import (
 from mttl.models.lightning.expert_module import ExpertModule, MoEModule
 from mttl.models.lightning.loggers import get_pl_loggers
 from mttl.models.monitors import get_monitors
-from mttl.utils import generate_random_string, rank_zero_only_and_wait, remote_login
+from mttl.utils import (
+    create_library,
+    generate_random_string,
+    remote_login,
+    upload_library,
+)
 
 
 def setup_profiler(args: ExpertConfig):
@@ -55,16 +59,6 @@ def train_experts(args: Args, model_class: Type[ExpertModule]):
     remote_login(args.remote_token)
     expert_library = None
     if args.library_id:
-
-        @rank_zero_only_and_wait(before=False, after=True)
-        def create_library(args):
-            expert_library = ExpertLibrary.get_expert_library(
-                repo_id=args.library_id,
-                create=True,
-                destination_id=args.destination_library_id,
-            )
-            return expert_library
-
         expert_library = create_library(args)
 
     loggers = get_pl_loggers(args)
@@ -190,28 +184,6 @@ def train_experts(args: Args, model_class: Type[ExpertModule]):
 
         module.load_state_dict(checkpoint)
         trainer.test(module, dm)
-
-        @rank_zero_only_and_wait(before=False, after=True)
-        def upload_library(expert_library, module):
-            if expert_library is not None:
-                # refresh expert library: so we dont overwrite the readme if the remote has changed.
-                expert_library.refresh_from_remote()
-
-                if isinstance(module, MoEModule):
-                    with expert_library.batched_commit():
-                        for expert_name in module.experts_names:
-                            expert = module.get_expert_instance(expert_name)
-                            expert_library.add_expert(expert, expert_name)
-                elif isinstance(module, ExpertModule):
-                    expert = module.as_expert()
-                    expert_name = (
-                        args.expert_name
-                        or args.finetune_task_name
-                        or generate_random_string()
-                    )
-                    expert_library.add_expert(expert, expert_name)
-                else:
-                    raise ValueError("Model class not recognized")
 
         upload_library(expert_library, module)
